DTOs, service done, Facade and tests TODO, some moving around in packages because packages felt incorrect
