--- lost-and-found-service-api/src/main/java/cz.muni.fi/dto/CategoryDto.java
@@ -1,4 +1,0 @@
-package cz.muni.fi.dto;
-
-public class CategoryDto {
-}
