--- lost-and-found-service-api/src/main/java/cz.muni.fi/dto/ItemDto.java
@@ -1,4 +1,0 @@
-package cz.muni.fi.dto;
-
-public class ItemDto {
-}
