+++ lost-and-found-service-api/src/main/java/cz/muni/fi/dto/ItemDTO.java
@@ -1,0 +1,4 @@
+package cz.muni.fi.dto;
+
+public class ItemDTO {
+}
