+++ lost-and-found-service-api/src/main/java/cz/muni/fi/enums/Status.java
