+++ lost-and-found-service-api/src/main/java/cz/muni/fi/facade/CategoryFacade.java
@@ -1,0 +1,20 @@
+package cz.muni.fi.facade;
+
+import cz.muni.fi.dto.CategoryDTO;
+
+import java.util.List;
+
+public interface CategoryFacade {
+
+    CategoryDTO findCategoryById(Long userId);
+
+    /**
+     * Create a given category
+     */
+    void createCategory(CategoryDTO c);
+
+    /**
+     * Get all categories
+     */
+    List<CategoryDTO> getAllCategories();
+}
