--- lost-and-found-service-persistence/src/main/java/cz/muni/fi/PersistenceApplicationContext.java
@@ -1,63 +1,0 @@
-package cz.muni.fi;
-
-import org.hibernate.jpa.HibernatePersistenceProvider;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.dao.annotation.PersistenceExceptionTranslationPostProcessor;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import org.springframework.instrument.classloading.InstrumentationLoadTimeWeaver;
-import org.springframework.instrument.classloading.LoadTimeWeaver;
-import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseBuilder;
-import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseType;
-import org.springframework.orm.jpa.JpaTransactionManager;
-import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
-
-import javax.sql.DataSource;
-
-/**
- * @author Terezia Slaninakova (445526)
- */
-@Configuration
-@EnableTransactionManagement
-@EnableJpaRepositories
-@ComponentScan(basePackages = "cz.muni.fi")
-public class PersistenceApplicationContext {
-
-    @Bean
-    public PersistenceExceptionTranslationPostProcessor postProcessor() {
-        return new PersistenceExceptionTranslationPostProcessor();
-    }
-
-    @Bean
-    public JpaTransactionManager transactionManager() {
-        return new JpaTransactionManager(entityManagerFactory().getObject());
-    }
-
-    @Bean
-    public LocalContainerEntityManagerFactoryBean entityManagerFactory() {
-        LocalContainerEntityManagerFactoryBean jpaFactoryBean = new LocalContainerEntityManagerFactoryBean();
-        jpaFactoryBean.setDataSource(db());
-        jpaFactoryBean.setLoadTimeWeaver(instrumentationLoadTimeWeaver());
-        jpaFactoryBean.setPersistenceProviderClass(HibernatePersistenceProvider.class);
-        return jpaFactoryBean;
-    }
-
-    @Bean
-    public LocalValidatorFactoryBean localValidatorFactoryBean() {
-        return new LocalValidatorFactoryBean();
-    }
-
-    @Bean
-    public LoadTimeWeaver instrumentationLoadTimeWeaver() {
-        return new InstrumentationLoadTimeWeaver();
-    }
-
-    @Bean
-    public DataSource db() {
-        EmbeddedDatabaseBuilder builder = new EmbeddedDatabaseBuilder();
-        return builder.setType(EmbeddedDatabaseType.DERBY).build();
-    }
-}
