+++ lost-and-found-service-persistence/src/main/java/cz/muni/fi/persistence/dao/ItemDaoImpl.java
@@ -1,0 +1,72 @@
+package cz.muni.fi.persistence.dao;
+
+import cz.muni.fi.persistence.entity.Item;
+import org.springframework.stereotype.Repository;
+
+import javax.persistence.EntityExistsException;
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+import java.util.List;
+
+/**
+ * @author Augustin Nemec
+ */
+@Repository
+public class ItemDaoImpl implements ItemDao {
+
+    @PersistenceContext
+    private EntityManager em;
+
+    @Override
+    public Item addItem(Item item) {
+        if (item == null) {
+            throw new IllegalArgumentException("Item is null");
+        }
+        try {
+            em.persist(item);
+        } catch (EntityExistsException e) {
+            throw new IllegalArgumentException("Item already exists");
+        }
+        return item;
+    }
+
+    @Override
+    public Item deleteItem(Item item) {
+        if (item == null || item.getId() == null) {
+            throw new IllegalArgumentException("Item is null");
+        }
+        try {
+            em.remove(item);
+        } catch (IllegalArgumentException e) {
+            throw new IllegalArgumentException("Nothing to remove");
+        }
+        return item;
+    }
+
+    @Override
+    public Item getItembyId(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("Item's id is null");
+        }
+        return em.find(Item.class, id);
+    }
+
+    @Override
+    public List<Item> getAllItems() {
+        return em.createQuery("select i from Item i", Item.class)
+                .getResultList();
+    }
+
+    @Override
+    public Item updateItem(Item item) {
+        if (item == null || item.getId() == null) {
+            throw new IllegalArgumentException("Item is null");
+        }
+        try {
+            em.merge(item);
+        } catch (IllegalArgumentException e) {
+            throw new IllegalArgumentException("Item does not exist");
+        }
+        return item;
+    }
+}
