+++ lost-and-found-service-persistence/src/main/java/cz/muni/fi/persistence/enums/Status.java
@@ -1,0 +1,13 @@
+package cz.muni.fi.persistence.enums;
+
+/**
+ *
+ * @author Augustin Nemec
+ */
+
+public enum Status {
+    CLAIM_RECEIVED,
+    IN_PROGRESS,
+    FOUND
+
+}
