+++ lost-and-found-service-service/src/main/java/cz/muni/fi/service/ItemServiceImpl.java
@@ -1,7 +1,7 @@
 package cz.muni.fi.service;
 
-import cz.muni.fi.dao.ItemDao;
-import cz.muni.fi.entity.Item;
+import cz.muni.fi.persistence.dao.ItemDao;
+import cz.muni.fi.persistence.entity.Item;
 import cz.muni.fi.service.exceptions.ServiceException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
